+++ CHANGELOG.md
@@ -8,3 +8,4 @@
 * Allow `Widget`s to register themselves
 * Fix `TextureWidget`s adding themselves to a `TextBuilder` multiple times
 * Fix render issues by using at least 1 splitting character
+* Add the `space` as a hardcoded 4px-wide whitespace character